backend/requirements.txt
# django
django
djangorestframework
django-filter
django-cors-headers
django-cleanup
drf-spectacular
drf-spectacular[sidecar]

# celery
django-celery-results
django-celery-beat
celery
celery[redis]

# servers + utils
redis
psycopg2-binary
sqlalchemy
sqlalchemy-utils
gunicorn
gevent
psutil
python-dotenv
markdown
authlib

# bsi
lark-parser
pyparsing
behave
pytest
pydantic
pyspellchecker
requests
pydot
numpy
rtree
networkx
pyyaml

# dev
django-debug-toolbar
flake8
deprecated
